make some structural changes
