+++ src/instructions/add.rs
@@ -1,4 +1,4 @@
-use crate::vm::VM;
+use crate::vm::vm::VM;
 
 use super::sign_extend;
 
@@ -30,7 +30,7 @@ impl VM {
 
 #[cfg(test)]
 mod tests {
-    use crate::vm::VM;
+    use super::*;
 
     #[test]
     fn test_op_add_register_mode() {
