+++ src/instructions/jmp.rs
@@ -1,4 +1,4 @@
-use crate::vm::VM;
+use crate::vm::vm::VM;
 
 impl VM {
     pub fn op_jmp(&mut self, instr: u16) {
