+++ src/instructions/ldi.rs
@@ -1,4 +1,4 @@
-use crate::vm::VM;
+use crate::vm::vm::VM;
 
 use super::sign_extend;
 
@@ -21,7 +21,7 @@ impl VM {
 
 #[cfg(test)]
 mod tests {
-    use crate::vm::VM;
+    use super::*;
 
     // Helper function to set up a VM instance for testing
     fn set_up_vm() -> VM {
