+++ src/lib.rs
@@ -1,7 +1,3 @@
-pub mod condition_flag;
 pub mod instructions;
-pub mod memory;
-pub mod opcode;
-pub mod registers;
 pub mod utils;
 pub mod vm;
