+++ src/main.rs
@@ -1,5 +1,5 @@
 use lc3_vm::utils::*;
-use lc3_vm::vm::VM;
+use lc3_vm::vm::vm::VM;
 fn main() {
     // Load Arguments
     load_arguments();
